J_T: Switching to new scripts
Jose_Testing scene: Removed Health and Enemy script on enemy fry. Replaced with nHealth and nEnemy on enemy fry. Replaced Damaging with nDamaging on onion weapon trigger on the player.

nEnemy: Added as a copy of the original Enemy script to be worked on. Now points to nHealth script.

nOnionWeapon: Now resets nDamaging attack trigger when needed.